Be more explicit about the type of initYCbCrConversion
Declare it as `void initYCbCrConversion(void)`, a nullary function in
all C standards, rather than as `void initYCbCrConversion()`, a nullary
function in C23 but a function of unspecified arguments in earlier
standards. This doesn’t fix anything (the code works the same in
practice in both cases) but is more explicit about intent.