Avoid function pointers of unspecified arguments (#53)
As of C23 (the default standard in GCC 15), `int foo()` is interpreted
as a nullary function like `int foo(void)`, as it would be in C++,
rather than as a function of unspecified arguments.

* Be more explicit about the type of initYCbCrConversion

Declare it as `void initYCbCrConversion(void)`, a nullary function in
all C standards, rather than as `void initYCbCrConversion()`, a nullary
function in C23 but a function of unspecified arguments in earlier
standards. This doesn’t fix anything (the code works the same in
practice in both cases) but is more explicit about intent.